post show

+++ database/migrations/2022_09_17_145236_create_comments_table.php
@@ -23,7 +23,7 @@ return new class extends Migration
             $table->boolean('approved')->default(0);
             $table->bigInteger('commentable_id');
             $table->string('commentable_type');
-            $table->integer('rate')->nullable();
+            $table->tinyInteger('rate')->nullable();
             $table->timestamps();
         });
     }
